refactor: rename internals to internal-constructs [sc-0]
