+++ packages/cli/src/constructs/check.ts
@@ -1,7 +1,7 @@
 import { AlertChannel } from './alert-channel'
 import { EnvironmentVariable } from './environment-variable'
 import { AlertChannelSubscription } from './alert-channel-subscription'
-import { Construct, Session, Ref } from '../internals'
+import { Construct, Session, Ref } from '../internal-constructs'
 import { CheckConfigDefaults } from '../services/checkly-config-loader'
 
 /**
