+++ packages/cli/src/internal-constructs/project.ts
